Calculate min coverage on execution
